Address code review findings on Admin/base and add PHPUnit tests
- Admin::add_to_runtime() coerces non-array filter values instead of
  hard-failing with a TypeError.
- Admin::__construct() drops the redundant $_GET isset check now that
  filter_input() is used directly, removing the phpcs:ignore.
- Admin::assets() falls back to empty dependencies/current version when
  the build manifest is missing, instead of leaving $asset undefined.
- base.php explicitly casts both operands before subtracting timestamps.
- tests/phpunit/bootstrap.php now loads the plugin via muplugins_loaded
  so plugin classes are available to PHPUnit tests.
- Add lightweight integration tests for the settings/caching REST
  controllers, the Admin asset fallback, and the NFD compat-check
  deactivation paths.

+++ inc/base.php
@@ -41,7 +41,7 @@ function web_set_plugin_install_date( string $value ): void {
  * @return int
  */
 function web_get_days_since_plugin_install_date(): int {
-	return absint( ( gmdate( 'U' ) - web_get_plugin_install_date() ) / DAY_IN_SECONDS );
+	return absint( ( (int) gmdate( 'U' ) - (int) web_get_plugin_install_date() ) / DAY_IN_SECONDS );
 }
 
 /**
